+++ CHANGELOG.md
@@ -9,6 +9,7 @@ __BEGIN_UNRELEASED__
 ## [Unreleased]
 ### Added
 
+- Added support for Event Visualizer on iOS.
 - Added iOS-only option to enable Heap SDK debug logging in the iOS console, enabled by default on dev builds.
 - Disabled tracking of view controllers on iOS when `enableNativeTouchEventCapture` is off.