+++ CHANGELOG.md
@@ -15,6 +15,19 @@ __BEGIN_UNRELEASED__
 ### Security
 __END_UNRELEASED__
 
+## [0.19.0] - 2022-04-12
+
+### Added
+
+- Added support for Event Visualizer on iOS, with a pairing option in the React Native developer menu.
+- Added iOS-only option to enable Heap SDK debug logging in the iOS console, enabled by default on dev builds.
+- Disabled tracking of view controllers on iOS when `enableNativeTouchEventCapture` is off.
+
+### Changed
+
+- Updated `Heap.setAppId` on iOS to turn off native touch capture on iOS.
+- Updated `Heap.setAppId` on iOS to enable logging on debug builds.
+
 ## [0.18.0] - 2022-02-28
 
 ### Fixed